:zap: :: update api

+++ app/src/main/java/com/narsha/sundolls_ep_android/data/network/dto/user/UserResponseDto.kt
@@ -6,6 +6,6 @@ data class UserResponseDto(
     val profileUrl: String,
     val schoolName: String,
     val tag: String,
-    val totalStudyTime: Int,
+    val totalStudyTime: Long,
     val username: String
 )
